+++ library/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mahendri <mahendri@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/08 15:28:49 by mahendri          #+#    #+#             */
-/*   Updated: 2026/01/22 22:56:46 by mahendri         ###   ########.fr       */
+/*   Updated: 2026/01/28 09:22:33 by mahendri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
