+++ library/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mahendri <mahendri@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/20 15:31:19 by mahendri          #+#    #+#             */
-/*   Updated: 2026/01/22 22:56:43 by mahendri         ###   ########.fr       */
+/*   Updated: 2026/01/28 09:17:27 by mahendri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && !(!s1[i] && !s2[i]))
